+++ handlers/gaming-session/delete-gaming-session.go
@@ -1,0 +1,37 @@
+package gaming_session
+
+import (
+	"github.com/bwmarrin/discordgo"
+)
+
+func DeleteGamingSession(s *discordgo.Session, i *discordgo.InteractionCreate) {
+	if mabarSession {
+		messageContent := "Buyar dulu kawanku, sampai jumpa di mabar berikutnya!👋"
+		err := s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+			Type: discordgo.InteractionResponseChannelMessageWithSource,
+			Data: &discordgo.InteractionResponseData{
+				Content: messageContent,
+			},
+		})
+
+		mabarSession = false
+		membersSession = nil
+
+		if err != nil {
+			panic(err)
+		}
+
+	} else {
+		messageContent := "❌ Tidak ada sesi mabar yang berjalan, kawan"
+		err := s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+			Type: discordgo.InteractionResponseChannelMessageWithSource,
+			Data: &discordgo.InteractionResponseData{
+				Content: messageContent,
+			},
+		})
+
+		if err != nil {
+			panic(err)
+		}
+	}
+}
