+++ handlers/gaming-session/util-gaming-session.go
@@ -1,0 +1,27 @@
+package gaming_session
+
+import (
+	"fmt"
+)
+
+var (
+	membersSession []string
+	mabarSession   bool
+)
+
+func GenerateMemberMention() string {
+	result := ""
+	for _, s := range membersSession {
+		result += fmt.Sprintf("<@%v>", s)
+	}
+	return result
+}
+
+func CheckJoin(userId string) bool {
+	for _, u := range membersSession {
+		if u == userId {
+			return true
+		}
+	}
+	return false
+}
